01-kafka-python/requirements.txt
confluent-kafka==1.9.2
Faker==15.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
fastavro==1.7.0